+++ app/forms/conversion/create_project_form.rb
@@ -8,6 +8,7 @@ class Conversion::CreateProjectForm < CreateProjectForm
   validates :provisional_conversion_date, presence: true
   validates :provisional_conversion_date, date_in_the_future: true, first_day_of_month: true
 
+  validates :urn, presence: true, existing_academy: true
   validate :urn_unique_for_in_progress_conversions, if: -> { urn.present? }
 
   validates :directive_academy_order, :assigned_to_regional_caseworker_team, inclusion: {in: [true, false]}